Add OperationQueue for managing asynchronous operations in state management
