+++ packages/protocol/src/state/State.ts
@@ -40,17 +40,6 @@ export class WithStateServiceProvider {
  * Utilities for runtime module state, such as get/set
  */
 export class State<Value> extends Mixin(WithPath, WithStateServiceProvider) {
-  private queue: Promise<void> = Promise.resolve();
-
-  private queueOperation<T>(operation: () => Promise<T>): Promise<T> {
-    const result = this.queue.then(() => operation());
-    this.queue = result.then(
-      () => undefined,
-      () => undefined
-    );
-    return result;
-  }
-
   /**
    * Creates a new state wrapper for the provided value type.
    *
@@ -142,19 +131,21 @@ export class State<Value> extends Mixin(WithPath, WithStateServiceProvider) {
    * @returns Option representation of the current state.
    */
   public async get(): Promise<Option<Value>> {
-    return await this.queueOperation(async () => {
-      const option = await this.witnessFromState();
+    return await container
+      .resolve(RuntimeMethodExecutionContext)
+      .operationQueue.queueOperation(async () => {
+        const option = await this.witnessFromState();
 
-      this.hasPathOrFail();
+        this.hasPathOrFail();
 
-      const stateTransition = StateTransition.from(this.path, option);
+        const stateTransition = StateTransition.from(this.path, option);
 
-      container
-        .resolve(RuntimeMethodExecutionContext)
-        .addStateTransition(stateTransition);
+        container
+          .resolve(RuntimeMethodExecutionContext)
+          .addStateTransition(stateTransition);
 
-      return option;
-    });
+        return option;
+      });
   }
 
   /**
@@ -169,22 +160,24 @@ export class State<Value> extends Mixin(WithPath, WithStateServiceProvider) {
    * @param value - Value to be set as the current state
    */
   public async set(value: Value) {
-    await this.queueOperation(async () => {
-      // link the transition to the current state
-      const fromOption = await this.witnessFromState();
-      const toOption = Option.fromValue(value, this.valueType);
+    await container
+      .resolve(RuntimeMethodExecutionContext)
+      .operationQueue.queueOperation(async () => {
+        // link the transition to the current state
+        const fromOption = await this.witnessFromState();
+        const toOption = Option.fromValue(value, this.valueType);
 
-      this.hasPathOrFail();
+        this.hasPathOrFail();
 
-      const stateTransition = StateTransition.fromTo(
-        this.path,
-        fromOption,
-        toOption
-      );
+        const stateTransition = StateTransition.fromTo(
+          this.path,
+          fromOption,
+          toOption
+        );
 
-      container
-        .resolve(RuntimeMethodExecutionContext)
-        .addStateTransition(stateTransition);
-    });
+        container
+          .resolve(RuntimeMethodExecutionContext)
+          .addStateTransition(stateTransition);
+      });
   }
 }
